Move classes for Unnest Operator to a separate package
Cherry-pick of https://github.com/prestosql/presto/pull/901 (https://github.com/prestosql/presto/pull/901)
