+++ presto-main/src/main/java/com/facebook/presto/operator/unnest/ArrayOfRowsUnnester.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.operator;
+package com.facebook.presto.operator.unnest;
 
 import com.facebook.presto.spi.PageBuilder;
 import com.facebook.presto.spi.block.Block;
