+++ presto-main/src/test/java/com/facebook/presto/operator/unnest/TestUnnestOperator.java
@@ -11,9 +11,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.operator;
+package com.facebook.presto.operator.unnest;
 
 import com.facebook.presto.metadata.MetadataManager;
+import com.facebook.presto.operator.DriverContext;
+import com.facebook.presto.operator.OperatorFactory;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.plan.PlanNodeId;
 import com.facebook.presto.spi.type.ArrayType;
